+++ app/backend/rota/volunteer_rota_summary.py
@@ -46,6 +46,9 @@ def get_summary_list_of_roles_and_groups_for_event_as_pd_df(
     single_df = from_list_of_day_summaries_to_single_df(
         all_day_summaries=all_day_summaries, event=event
     )
+    if len(single_df)==0:
+        return single_df
+
     single_df.loc['TOTAL'] = single_df.sum(axis=0,numeric_only=True)
     return single_df
 
@@ -145,6 +148,9 @@ def get_summary_list_of_teams_and_groups_for_events_as_pd_df(
     single_df = from_list_of_day_summaries_to_single_df(
         all_day_summaries=all_day_summaries, event=event
     )
+    if len(single_df)==0:
+        return single_df
+
     single_df.loc['TOTAL'] = single_df.sum(axis=0,numeric_only=True)
 
     return single_df